fix(build): fix signed/unsigned conversion errors on Linux
- Use std::underlying_type_t<IoEvent> for enum bitmask operations
- Change total_n from ssize_t to size_t in async_send/async_sendto
- Add missing <optional> includes for Linux compatibility
- Fix port vector to use unsigned short instead of int

+++ tests/net/test_tcp_transport_comprehensive.cpp
@@ -440,7 +440,7 @@ void test_concurrent_acceptor_operations() {
 
     // Create multiple acceptors sharing the same event loop
     std::vector<std::unique_ptr<Acceptor>> acceptors;
-    std::vector<int> ports;
+    std::vector<unsigned short> ports;
 
     for (int i = 0; i < 3; ++i) {
         auto acceptor = std::make_unique<Acceptor>(&loop);
